refactor: useWaitForImagesLoadファイルをts化 #111
